refactor(test): migrate email and phone helpers to three-tier architecture
- Create db-test-seeders/email.ts (2 seeders with @why-db-direct)
- Create db-test-assertions/email.ts (3 read-only assertions)
- Create db-test-seeders/phone.ts (3 seeders with @why-db-direct)
- Create db-test-assertions/phone.ts (1 read-only assertion)
- Delete api-test-helpers/email.ts and phone.ts
- Remove barrel re-exports from api-test-helpers/index.ts
- Eliminate generateTestReplyToken wrapper — consumers import
  generateReplyToken directly from source module
- Add initServices() to all functions accessing globalThis.services.db
- Update 15 consumer test files to import from correct tier

Closes #9360
