+++ turbo/apps/web/app/api/webhooks/clerk/__tests__/route.test.ts
@@ -13,7 +13,6 @@ import {
   createTestSecret,
   createTestVariable,
   createTestAgentSession,
-  createTestEmailThreadSession,
   insertOrgDefaultModelProvider,
   insertTestStorage,
   insertTestStorageVersion,
@@ -37,6 +36,7 @@ import {
   insertUserCacheEntry,
   insertOrgCacheEntry,
 } from "../../../../../src/__tests__/api-test-helpers";
+import { createTestEmailThreadSession } from "../../../../../src/__tests__/db-test-seeders/email";
 import {
   insertTestSlackOrgInstallation,
   insertTestSlackOrgConnection,
