+++ turbo/apps/web/app/api/zero/developer-support/__tests__/route.test.ts
@@ -6,8 +6,8 @@ import {
   createTestRequest,
   createTestCompose,
   insertOrgMembersCacheEntry,
-  findTestOutboxItems,
 } from "../../../../../src/__tests__/api-test-helpers";
+import { findTestOutboxItems } from "../../../../../src/__tests__/db-test-assertions/email";
 import {
   testContext,
   uniqueId,
