+++ turbo/apps/web/app/api/zero/phone/setup/__tests__/route.test.ts
@@ -7,10 +7,8 @@ import {
 } from "../../../../../../src/__tests__/api-test-helpers";
 import { testContext } from "../../../../../../src/__tests__/test-helpers";
 import { mockClerk } from "../../../../../../src/__tests__/clerk-mock";
-import {
-  createPhoneOrg,
-  getOrgAgentphoneConfig,
-} from "../../../../../../src/__tests__/api-test-helpers/phone";
+import { createPhoneOrg } from "../../../../../../src/__tests__/db-test-seeders/phone";
+import { getOrgAgentphoneConfig } from "../../../../../../src/__tests__/db-test-assertions/phone";
 import { server } from "../../../../../../src/mocks/server";
 import { reloadEnv } from "../../../../../../src/env";
