+++ turbo/apps/web/app/api/zero/report-error/__tests__/route.test.ts
@@ -7,8 +7,8 @@ import {
   createTestRequest,
   createTestOrg,
   createTestCompose,
-  findTestOutboxItems,
 } from "../../../../../src/__tests__/api-test-helpers";
+import { findTestOutboxItems } from "../../../../../src/__tests__/db-test-assertions/email";
 import {
   testContext,
   uniqueId,
