+++ turbo/apps/web/src/__tests__/api-test-helpers/index.ts
@@ -10,10 +10,8 @@ export * from "./storage";
 export * from "./secrets";
 export * from "./connectors";
 export * from "./callbacks";
-export * from "./email";
 export * from "./github";
 export * from "./telegram";
-export * from "./phone";
 export * from "./credits";
 export * from "./users";
 export * from "./skills";
